Feat: refect app_page components
