+++ src/app/(private)/app/subjects/topics/page.tsx
@@ -1,10 +1,10 @@
-import DashboardHeader from '@/components/app_page/dashboard/DashboardHeader';
+import Header from '@/components/app_page/Header';
 import { SubjectTopicsContent } from '@/components/subjects_page/SubjectTopicsContent';
 
 export default function Topics() {
   return (
     <div className="min-h-screen bg-slate-50 flex flex-col">
-      <DashboardHeader />
+      <Header />
       <main className="flex-grow pb-6">
         <SubjectTopicsContent />
       </main>
