+++ src/app/(private)/app/topics/contents/page.tsx
@@ -1,10 +1,10 @@
-import DashboardHeader from '@/components/app_page/dashboard/DashboardHeader';
+import Header from '@/components/app_page/Header';
 import { TopicContentPageContent } from '@/components/topics_page/TopicContentPageContent';
 
 export default function TopicContentsPage() {
   return (
     <div className="min-h-screen bg-slate-50 flex flex-col">
-      <DashboardHeader />
+      <Header />
       <main className="flex-grow pb-6">
         <TopicContentPageContent />
       </main>
